separate add item function
